Introduce richer Tetromino struct
The new Tetromino struct encapsulates a tetromino's type,
the number of rotation variants for this type, and its
initial unrotated cell layout.

This makes a lot more sense for the rotation system. It
now has all the required information in a single place.

Additionally this commit fixes some erroneous rotation
test data.